Update MongoScheduler's sleep period
Substitute `self.sync_every` with `self.max_interval`. The former
is set to 180 seconds, while the latter defaults to 5 seconds and
it's configurable via celery's settings. This change effectively
makes the scheduler service sleep for a shorter period of time in
between checking whether schedules are due. Otherwise, a schedule
with a frequency of 2 minutes, for instance, could be blocked for
up to 3 minutes seconds, if the former schedule had a smaller interval.

Hopefully, this change will allow tasks to be scheduled "more" on
time.